add build.zig, cross-platform builds and no Visual Studio dependency
This adds a build.zig file which enables building this project with
a single dependency, zig.  It's a 75 MB download from
https://ziglang.org/download.

This removes needing to install Visual Studio which clang++ depends on.
You can confirm this by trying to build this project inside
"Windows Sandbox".

This also enables cross-compiling from other platforms (i.e. linux/macos).
I compiled this on linux for windows via:

    zig build -Dtarget=x86_64-windows

And then I was able to run it on linux with wine64.